Restructure Directories
Without this change, `ember-formulaic` was imported via:

```js
// tests/test-helper.js

import './test-helpers';
```

This is confusing as a consumer of the library, since nothing is
`ember-formulaic` specific.

In addition to the confusion, anything else declared in the addon's
`test-support/helpers` directory is directly imported to the consumer
app's `tests/helpers` directory.

This commit limits the exposure of addon internals. Instead of exposing
everything in the addon, the only file exposed is
`test-support/helpers/ember-formulaic/test-helpers.js`.

The side-effects of this change are breaking, but minimal. Consumers
must now import the addon like so:

```js
// tests/test-helper.js

import './ember-formulaic/test-helpers';
```

Although this is a breaking change, the previous usage was never
properly documented.